gh-95: Update to Swagger v0.2.4

+++ models/thing_templates_list_response.go
@@ -35,7 +35,7 @@ type ThingTemplatesListResponse struct {
 	ThingTemplates []*ThingTemplateGetResponse `json:"thingTemplates"`
 
 	// The total number of model manifests for the query. The number of items in a response may be smaller due to paging.
-	TotalResults int32 `json:"totalResults,omitempty"`
+	TotalResults int64 `json:"totalResults,omitempty"`
 }
 
 // Validate validates this thing templates list response
